Fixing tests
For some reason I cannot run the tests locally right now, but I think this should do the job/].
