+++ tests/test_level_01.py
@@ -1,7 +1,7 @@
 import hedy
 import textwrap
 from Tester import HedyTester
-from hedy import local_keywords_enabled
+from test_translating import check_local_lang_bool
 
 
 class TestsLevel1(HedyTester):
@@ -92,21 +92,23 @@ class TestsLevel1(HedyTester):
 
     self.single_level_tester(code=code, expected=expected)
     
-  if local_keywords_enabled:
-    def test_print_dutch(self):
-      result = hedy.transpile("print Hallo welkom bij Hedy!", self.level, lang="nl")
-      expected = "print('Hallo welkom bij Hedy!')"
-      self.assertEqual(expected, result.code)
-      self.assertEqual(False, result.has_turtle)
-      self.assertEqual('Hallo welkom bij Hedy!', HedyTester.run_code(result))
-    def test_ask_dutch_error(self):
-      code = textwrap.dedent("""ask Heb je er zin?""")
+  @check_local_lang_bool
+  def test_print_dutch(self):
+    result = hedy.transpile("print Hallo welkom bij Hedy!", self.level, lang="nl")
+    expected = "print('Hallo welkom bij Hedy!')"
+    self.assertEqual(expected, result.code)
+    self.assertEqual(False, result.has_turtle)
+    self.assertEqual('Hallo welkom bij Hedy!', HedyTester.run_code(result))
+  
+  @check_local_lang_bool
+  def test_ask_dutch_error(self):
+    code = textwrap.dedent("""ask Heb je er zin?""")
 
-      with self.assertRaises(hedy.exceptions.ParseException) as context:
-        result = hedy.transpile(code, self.level, lang="nl")
-      self.assertEqual('Parse', context.exception.error_code)
-      self.assertEqual(1, context.exception.error_location[0])
-      self.assertEqual('?', context.exception.error_location[1])
+    with self.assertRaises(hedy.exceptions.ParseException) as context:
+      result = hedy.transpile(code, self.level, lang="nl")
+    self.assertEqual('Parse', context.exception.error_code)
+    self.assertEqual(1, context.exception.error_location[0])
+    self.assertEqual('?', context.exception.error_location[1])
 
   # ask tests
   def test_ask(self):
