+++ tests/test_level_10.py
@@ -1,7 +1,8 @@
 import hedy
 import textwrap
 from test_level_01 import HedyTester
-from hedy import local_keywords_enabled
+from test_translating import check_local_lang_bool
+
 class TestsLevel10(HedyTester):
   level = 10
   
@@ -39,18 +40,19 @@ class TestsLevel10(HedyTester):
       print(f'{shark} shark')""")
 
     self.assertEqual(expected, result.code)
-  if local_keywords_enabled: 
-    def test_for_list_dutch(self):
-      code = textwrap.dedent("""\
-      dieren is hond, kat, papegaai
-      voor dier in dieren
-        print dier""")
+    
+  @check_local_lang_bool
+  def test_for_list_dutch(self):
+    code = textwrap.dedent("""\
+    dieren is hond, kat, papegaai
+    voor dier in dieren
+      print dier""")
 
-      result = hedy.transpile(code, self.level, lang="nl")
+    result = hedy.transpile(code, self.level, lang="nl")
 
-      expected = textwrap.dedent("""\
-      dieren = ['hond', 'kat', 'papegaai']
-      for dier in dieren:
-        print(f'{dier}')""")
+    expected = textwrap.dedent("""\
+    dieren = ['hond', 'kat', 'papegaai']
+    for dier in dieren:
+      print(f'{dier}')""")
 
-      self.assertEqual(expected, result.code)
+    self.assertEqual(expected, result.code)
