+++ tests/test_translating.py
@@ -1,7 +1,17 @@
 import unittest
+import hedy 
 import utils
 from website import translating
 
+def check_local_lang_bool(func):
+    def inner(self):
+        if not hedy.local_keywords_enabled:
+            return
+
+        return func(self)
+
+    return inner
+  
 class TestTranslating(unittest.TestCase):
   def test_change_nested_records(self):
     data = {}
